+++ hail/python/hail/plot2/__init__.py
@@ -1,4 +1,6 @@
-from .ggplot import aes, ggplot, geom_point, geom_line
+from .ggplot import ggplot
+from .aes import aes
+from .geoms import geom_line, geom_point
 
 
 __all__ = [
